// src/app/types/SpecificEquipment.ts
export interface SpecificEquipment {
  id: string;
  description: string; // For "Description"
  scope: string; // For "Scope"
  tagID: string; // For "Tag Id"
  make: string; // For "Make"
  model: string; // For "Model"
  serialNumber: string; // For "Serial Number"
  type: string; // For "Type"
  range: string; // For "Range"
  certificateNo: string; // For "Certificate No."
  traceability: string; // For "Traceability"
}